Implemented SVN feature request 1664472: Add support for setting the event properties
Added the possibility the pass in the property name/value pairs to FireEvent method. These will be set on the event object so that any (jscript) eventhandler attached to the event can make use if these values.

Also created overloads for KeyDown, KeyPress, KeyUp. They accept a char which is pressed (will be set on the event object using the new FireEvent functionality).

TextField.TypeText will now set the keyCode property of the event object during KeyUp, KeyPress, KeyDown (making use of the new overloads).