Fixed some bugs in IsElementReferenceStillValid in both IE and FireFox.
Refactored the NativeElement updating process so that it automatically checks whether the reference is valid and refreshes it if needed otherwise uses the cached element.  Cleaned up the logic to remove a bunch of special cases that previously caused unnecessary refreshes to occur.

Modified TryFuncUntilTimeOut to use an adaptive polling delay starting with 1ms and doubling until the SleepTime is reached.

Started using TimeSpan to denote timespans for some internal classes.  Public API still uses integer seconds or milliseconds as before (but we should consider revisiting this because it is confusing to have timespans in different units in different places).

Fixed a couple of incorrect namespace names for some internal types (this accounts for most of the changed files).

Partially reverted some changes to the Core.csproj.  Looks like it got a bad App.config added to it and the System.Core reference vanished.  The way this is configured now is based on the current VS configuration.  (ie. Debug .Net 2.0 / Debug .Net 3.5)

Note: I am observing a failure in the TextFieldTests on IE 8 because while the KeyPress event does get delivered to the browser, the keyCode is always set to 0.  I don't think this is caused by my changes here but we will need to investigate it further.